edited configuration paths
Former-commit-id: 6750f87249ef776000717ac9759a4adaf3ae1765

+++ python/postprocessing/visualization/api/common/plotters.py
@@ -44,11 +44,12 @@ class MultiLineMatplotlibPlotter(Plotter):
         fig.show()
 
     def save(self, file_name):
-        fig = self._build()
-
         file_path = os.path.join(self.output_dir, file_name + self.ext)
-        fig.savefig(file_path)
-        log.info('%s visualisation saved in %s', file_name, file_path)
+        if not os.path.isfile(file_path):
+            fig = self._build()
+            fig.savefig(file_path)
+            log.debug('%s visualisation saved in %s', file_name, file_path)
+        return file_path
 
     def _build(self):
         fig, axs = plt.subplots(nrows=self.n_rows, ncols=self.n_cols, sharex=True, sharey=True, figsize=self.fig_size)
@@ -112,11 +113,12 @@ class MultiWaterfallMatplotlibPlotter(Plotter):
         fig.show()
 
     def save(self, file_name):
-        fig = self._build()
-
         file_path = os.path.join(self.output_dir, file_name + self.ext)
-        fig.savefig(file_path)
-        log.debug('%s visualisation saved in %s', file_name, file_path)
+        if not os.path.isfile(file_path):
+            fig = self._build()
+            fig.savefig(file_path)
+            log.debug('%s visualisation saved in %s', file_name, file_path)
+        return file_path
 
     def _build(self):
         fig, axs = plt.subplots(nrows=self.n_rows, ncols=self.n_cols, sharex=True, sharey=True, figsize=self.fig_size)
